increased loading speed

+++ workers/tracker/src/main.py
@@ -9,7 +9,8 @@ from ml_core import AnnotationItem
 from src.annotate_json import tracking
 from src.config import settings
 from src.dataset_export.config import SAVE_FOLDER
-from src.services.storage import YandexDiskStorage
+from src.services.redis import RedisService
+from src.services.storage import YndexDiskStorage
 
 broker = RabbitBroker(
     settings.RABBITMQ_URL,
@@ -24,20 +25,19 @@ queue = RabbitQueue(
 
 @broker.subscriber(queue)
 async def handle_message(data: dict, msg: RabbitMessage):
-    storage = YandexDiskStorage(settings.YANDEX_DISK_TOKEN)
+    storage = YndexDiskStorage(settings.YANDEX_DISK_TOKEN)
+    redis = RedisService(settings.REDIS_URL, 600)
+    path_file = str(SAVE_FOLDER / 'video.mp4')
+    path_zip = None
     try:
         task_id = data['task_id']
         storage_path = data['storage_path']
-        metadata = data['metadata']
-        metadata = json.loads(metadata)
+        metadata = json.loads(data['metadata'])
         export_type = metadata['type']
         items: list[AnnotationItem] = [
             AnnotationItem(**x) for x in metadata['prompt']
         ]
-        print(
-            f'storage_path={storage_path}, type={export_type}, metadata={items}',
-        )
-        path_file = str(SAVE_FOLDER / 'video.mp4')
+        await redis.set_status(task_id, 'processing')
         await storage.download(storage_path, path_file)
 
         path_zip = await asyncio.to_thread(
@@ -50,6 +50,7 @@ async def handle_message(data: dict, msg: RabbitMessage):
         print(path_zip)
 
         await storage.upload(f'video/{task_id}.zip', path_zip)
+        await redis.set_status(task_id, 'done')
         await msg.ack()
     except Exception as e:
         print('ERROR:', repr(e))
@@ -57,4 +58,5 @@ async def handle_message(data: dict, msg: RabbitMessage):
     finally:
         await storage.close()
         Path(path_file).unlink(missing_ok=True)
-        Path(path_zip).unlink(missing_ok=True)
+        if path_zip:
+            Path(path_zip).unlink(missing_ok=True)
